requirements.txt
pytest~=8.3.3
fastapi~=0.115.5
sentry-sdk~=2.19.0
pandas~=2.2.3
SQLAlchemy~=2.0.36
dbfread~=2.0.7
pydantic~=2.10.2
python-dotenv~=1.0.1
uvicorn==0.34.0
httpx==0.28.1